# backend/openapi-schema.yaml
openapi: 3.0.3
info:
  title: Poultry Hero API
  version: 0.1.0
  description: SaaS Poultry Trading ERP — backend foundation (Phase 0/1).
paths:
...
  /api/v1/tenant/quotations/{id}/convert-to-sales/:
    post:
      operationId: tenant_quotations_convert_to_sales_create
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this quotation.
        required: true
      tags:
      - tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationDetail'
          description: ''
  /api/v1/tenant/quotations/{id}/lines/:
    get:
      operationId: tenant_quotations_lines_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this quotation.
        required: true
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationDetail'
          description: ''
    post:
      operationId: tenant_quotations_lines_create
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this quotation.
        required: true
      tags:
      - tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationDetail'
          description: ''
...
  /api/v1/tenant/quotations/{id}/print-preview/:
    get:
      operationId: tenant_quotations_print_preview_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this quotation.
        required: true
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationDetail'
          description: ''
...
  /api/v1/tenant/quotations/{id}/stock-warning/:
    get:
      operationId: tenant_quotations_stock_warning_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this quotation.
        required: true
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationDetail'
          description: ''
  /api/v1/tenant/quotations/expire-overdue/:
    post:
      operationId: tenant_quotations_expire_overdue_create
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      tags:
      - tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/QuotationDetail'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationDetail'
          description: ''
  /api/v1/tenant/quotations/price-preview/:
    get:
      operationId: tenant_quotations_price_preview_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationDetail'
          description: ''
...
components:
  schemas:
...
    LineTypeEnum:
      enum:
      - product
      - by_product
      - service
      - other
      type: string
      description: |-
        * `product` - Product
        * `by_product` - By-product
        * `service` - Service
        * `other` - Other
...
    PriceSourceEnum:
      enum:
      - default_product_price
      - customer_special_price
      - manual_override
      - free_product
      type: string
      description: |-
        * `default_product_price` - Default Product Price
        * `customer_special_price` - Customer Special Price
        * `manual_override` - Manual Override
        * `free_product` - Free Product
...
    QuotationDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        quotation_number:
          type: string
          maxLength: 64
        customer:
          type: integer
        customer_name_snapshot:
          type: string
          maxLength: 255
        customer_trn_snapshot:
          type: string
          maxLength: 32
        customer_phone_snapshot:
          type: string
          maxLength: 32
        customer_address_snapshot:
          type: string
        quotation_date:
          type: string
          format: date
        valid_until:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/Status50fEnum'
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        discount_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        taxable_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        vat_rate:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        vat_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        terms_and_conditions:
          type: string
        notes:
          type: string
        internal_notes:
          type: string
        sent_at:
          type: string
          format: date-time
          nullable: true
        accepted_at:
          type: string
          format: date-time
          nullable: true
        rejected_at:
          type: string
          format: date-time
          nullable: true
        expired_at:
          type: string
          format: date-time
          nullable: true
        converted_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        converted_sales_invoice:
          type: integer
          nullable: true
        cancel_reason:
          type: string
        reject_reason:
          type: string
        created_by:
          type: integer
          nullable: true
        updated_by:
          type: integer
          nullable: true
        sent_by:
          type: integer
          nullable: true
        accepted_by:
          type: integer
          nullable: true
        rejected_by:
          type: integer
          nullable: true
        converted_by:
          type: integer
          nullable: true
        cancelled_by:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/QuotationLine'
          readOnly: true
        status_history:
          type: array
          items:
            $ref: '#/components/schemas/QuotationStatusHistory'
          readOnly: true
      required:
      - created_at
      - customer
      - id
      - lines
      - quotation_date
      - quotation_number
      - status_history
      - updated_at
      - valid_until
    QuotationLine:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        product:
          type: integer
          nullable: true
        product_name_snapshot:
          type: string
          readOnly: true
        product_sku_snapshot:
          type: string
          readOnly: true
        line_type:
          $ref: '#/components/schemas/LineTypeEnum'
        quantity_cartons:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        quantity_pieces:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        quantity_kg:
          type: string
          format: decimal
          pattern: ^-?\d{0,11}(?:\.\d{0,3})?$
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        price_type:
          type: string
          maxLength: 16
        price_source:
          $ref: '#/components/schemas/PriceSourceEnum'
        is_free:
          type: boolean
        free_reason:
          type: string
        line_subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        taxable_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        vat_rate:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        vat_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        line_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        notes:
          type: string
        sort_order:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
        is_stock_tracked:
          type: boolean
          readOnly: true
      required:
      - id
      - is_stock_tracked
      - line_subtotal
      - line_total
      - product_name_snapshot
      - product_sku_snapshot
      - taxable_amount
      - vat_amount
...
    QuotationStatusHistory:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        from_status:
          type: string
          readOnly: true
        to_status:
          type: string
          readOnly: true
        reason:
          type: string
          readOnly: true
        changed_by:
          type: integer
          readOnly: true
          nullable: true
        changed_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - changed_at
      - changed_by
      - from_status
      - id
      - reason
      - to_status
...
    Status50fEnum:
      enum:
      - draft
      - sent
      - accepted
      - rejected
      - expired
      - converted
      - cancelled
      type: string
      description: |-
        * `draft` - Draft
        * `sent` - Sent
        * `accepted` - Accepted
        * `rejected` - Rejected
        * `expired` - Expired
        * `converted` - Converted
        * `cancelled` - Cancelled
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT